backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai==1.50.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
chromadb==0.5.15
httpx==0.27.2
langchain-huggingface==0.1.0
sentence-transformers==3.0.1